refactored to (rough) session based messaging
