+++ apollo/handlers/agent.py
@@ -1,4 +1,4 @@
-from uuid import UUID
+import uuid
 
 from fastapi import Depends, WebSocket
 from starlette.websockets import WebSocketDisconnect
@@ -7,7 +7,7 @@ from sqlalchemy.orm import Session
 from apollo.lib.router import SecureRouter
 from apollo.lib.schemas.agent import AgentSchema, CreateAgentSchema
 from apollo.lib.security import Allow, Authenticated
-from apollo.lib.websocket_manager import WebSocketManager
+from apollo.lib.websocket import WebSocketManager
 from apollo.models import get_session, save
 from apollo.models.agent import Agent
 from apollo.models.oauth import OAuthClient
@@ -29,11 +29,15 @@ def post_agent(agent_data: CreateAgentSchema,
 
 
 @router.websocket("/agent/{agent_id}/shell", permission='public')
-async def shell(websocket: WebSocket, agent_id: UUID):
+async def shell(websocket: WebSocket, agent_id: uuid.UUID):
     await websocket.accept()
+    session_id = uuid.uuid4()
+
+    manager = WebSocketManager()
+    manager.sessions[session_id] = websocket
     try:
         while True:
             command = await websocket.receive_text()
-            await WebSocketManager().send_message(agent_id, websocket, command)
+            await WebSocketManager().send_message(agent_id, session_id, command)
     except WebSocketDisconnect:
         return
