+++ apollo/lib/websocket/__init__.py
@@ -10,13 +10,15 @@ from apollo.lib.singleton import Singleton
 class WebSocketManager(metaclass=Singleton):
     def __init__(self):
         self.connections: Dict[uuid.UUID, WebSocket] = {}
-        self.messages: Dict[uuid.UUID, WebSocket] = {}
+        self.sessions: Dict[uuid.UUID, WebSocket] = {}
 
     async def _keep_connection_open(self, websocket: WebSocket):
         try:
             while True:
                 response = await websocket.receive_json()
-                requesting_socket = self.messages.pop(response['message_id'])
+                requesting_socket = self.connections[
+                    uuid.UUID(response['session_id'])
+                ]
                 await requesting_socket.send_text(response['message'])
         except WebSocketDisconnect:
             return
@@ -32,24 +34,6 @@ class WebSocketManager(metaclass=Singleton):
 
         raise error
 
-    async def _wait_for_response(self, target_websocket: WebSocket,
-                                 message_id: uuid.UUID):
-        try:
-            while True:
-                response = await target_websocket.receive_json()
-                try:
-                    if response['message_id'] == str(message_id):
-                        return response
-                except KeyError:
-                    continue
-        except WebSocketDisconnect:
-            return
-        except RuntimeError as e:
-            self._check_runtime_error(
-                error=e,
-                message='Cannot call "receive" once a disconnect'
-            )
-
     async def add_and_connect_websocket(self, client_id: uuid.UUID,
                                         websocket: WebSocket):
         self.connections[client_id] = websocket
@@ -58,16 +42,14 @@ class WebSocketManager(metaclass=Singleton):
         await self._keep_connection_open(websocket)
 
     async def send_message(
-        self, target_websocket_id: uuid.UUID, requesting_socket: WebSocket,
+        self, target_websocket_id: uuid.UUID, session_id: uuid.UUID,
         message: str
     ):
-        message_id = uuid.uuid4()
         target_websocket = self.connections[target_websocket_id]
-        self.messages[str(message_id)] = requesting_socket
         await self._send_message(
             target_websocket=target_websocket,
             message={
-                'message_id': str(message_id),
+                'session_id': str(session_id),
                 'message': message
             })
